requirements.txt
streamlit
openai
python-dotenv
Pillow
pytesseract
pyautogui
screeninfo
SpeechRecognition
gtts 